feat: cartItem 컴포넌트를 제거, API 명세서에 맞게 carItem의 id와 product의 id를 비동일시하도록 변경
